reorganization, cleanup and use settings_test
