+++ codecov_auth/signals.py
@@ -1,15 +1,10 @@
-import json
-import logging
 from typing import Any, Dict, Optional, Type
 
-from django.conf import settings
 from django.db.models.signals import post_save
 from django.dispatch import receiver
-from google.cloud import pubsub_v1
 
 from codecov_auth.models import OrganizationLevelToken, Owner, OwnerProfile
-
-log = logging.getLogger(__name__)
+from utils.shelter import ShelterPubsub
 
 
 @receiver(post_save, sender=Owner)
@@ -18,42 +13,6 @@ def create_owner_profile_when_owner_is_created(
 ) -> Optional[OwnerProfile]:
     if created:
         return OwnerProfile.objects.create(owner_id=instance.ownerid)
-
-
-class ShelterPubsub:
-    pubsub_publisher = None
-    _instance = None
-
-    @classmethod
-    def get_instance(cls) -> "ShelterPubsub":
-        """
-        This class needs the Django settings to be fully loaded before it can be instantiated,
-        therefore use this method to get an instance rather than instantiating directly.
-        """
-        if cls._instance is None:
-            cls._instance = cls()
-        return cls._instance
-
-    def __init__(self) -> None:
-        if not self.pubsub_publisher:
-            self.pubsub_publisher = pubsub_v1.PublisherClient()
-        pubsub_project_id: str = settings.SHELTER_PUBSUB_PROJECT_ID
-
-        # topic_id has REPO in the name but it is used for all types of objects
-        topic_id: str = settings.SHELTER_PUBSUB_SYNC_REPO_TOPIC_ID
-        self.topic_path = self.pubsub_publisher.topic_path(pubsub_project_id, topic_id)
-
-    def publish(self, data: Dict[str, Any]) -> None:
-        try:
-            self.pubsub_publisher.publish(
-                self.topic_path,
-                json.dumps(data).encode("utf-8"),
-            )
-        except Exception as e:
-            log.warning(
-                "Failed to publish a message",
-                extra=dict(data_to_publish=data, error=e),
-            )
 
 
 @receiver(
