+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-DJANGO_SETTINGS_MODULE = codecov.settings_dev
+DJANGO_SETTINGS_MODULE = codecov.settings_test
 addopts = -p no:warnings --ignore=shared --ignore-glob=**/test_results*
